# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Karaoke CT API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Musica {
  id: number;
  artista: string;
  musica: string;
  /** @nullable */
  inicio: string | null;
}

export interface MusicaInput {
  id: number;
  /** @minLength 1 */
  artista: string;
  /** @minLength 1 */
  musica: string;
  inicio?: string;
}

export interface MusicaUpdate {
  /** @minLength 1 */
  artista?: string;
  /** @minLength 1 */
  musica?: string;
  inicio?: string;
}

export interface MusicaSearchResult {
  data: Musica[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface CatalogStats {
  totalSongs: number;
  totalArtists: number;
}

export type KaraokeUserSubscriptionStatus = typeof KaraokeUserSubscriptionStatus[keyof typeof KaraokeUserSubscriptionStatus];


export const KaraokeUserSubscriptionStatus = {
  active: 'active',
  cancelled: 'cancelled',
  expired: 'expired',
} as const;

export interface KaraokeUser {
  id: number;
  email: string;
  /** @nullable */
  woocommerceCustomerId?: string | null;
  subscriptionStatus: KaraokeUserSubscriptionStatus;
  accessGranted: boolean;
  /** @nullable */
  accessGrantedAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}

export interface WooCommerceEvent {
  event: string;
  /** @nullable */
  customer_id?: string | null;
  /** @nullable */
  customer_email?: string | null;
  /** @nullable */
  subscription_status?: string | null;
  /** @nullable */
  order_id?: string | null;
}

export interface WebhookResult {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type SearchMusicasParams = {
/**
 * Search query
 */
q?: string;
/**
 * Page number (1-based)
 */
page?: number;
/**
 * Results per page
 */
limit?: number;
};