# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Karaoke CT API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: musicas
    description: Song catalog operations
  - name: users
    description: User management operations
  - name: webhook
    description: Webhook integrations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /musicas/search:
    get:
      operationId: searchMusicas
      tags: [musicas]
      summary: Search songs
      description: Real-time search across artista, musica, and inicio fields with pagination
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search query
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: Page number (1-based)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 24
          description: Results per page
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicaSearchResult"

  /musicas/stats:
    get:
      operationId: getMusicasStats
      tags: [musicas]
      summary: Catalog statistics
      description: Returns total songs count and unique artists count
      responses:
        "200":
          description: Catalog stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogStats"

  /musicas:
    post:
      operationId: createMusica
      tags: [musicas]
      summary: Add a song
      description: Add a new song to the catalog (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MusicaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Musica"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: ID already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /musicas/{id}:
    get:
      operationId: getMusica
      tags: [musicas]
      summary: Get a song by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Song found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Musica"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMusica
      tags: [musicas]
      summary: Update a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MusicaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Musica"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMusica
      tags: [musicas]
      summary: Delete a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List subscribers
      description: List all active subscribers (admin)
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KaraokeUser"

  /users/{id}/revoke:
    post:
      operationId: revokeUserAccess
      tags: [users]
      summary: Revoke user access
      description: Immediately revoke a user's karaoke access
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Access revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KaraokeUser"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /webhook/woocommerce:
    post:
      operationId: woocommerceWebhook
      tags: [webhook]
      summary: WooCommerce subscription webhook
      description: Handles subscription purchase, renewal, and cancellation events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WooCommerceEvent"
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Musica:
      type: object
      required: [id, artista, musica, inicio]
      properties:
        id:
          type: integer
        artista:
          type: string
        musica:
          type: string
        inicio:
          type: ["string", "null"]

    MusicaInput:
      type: object
      required: [id, artista, musica]
      properties:
        id:
          type: integer
        artista:
          type: string
          minLength: 1
        musica:
          type: string
          minLength: 1
        inicio:
          type: string

    MusicaUpdate:
      type: object
      properties:
        artista:
          type: string
          minLength: 1
        musica:
          type: string
          minLength: 1
        inicio:
          type: string

    MusicaSearchResult:
      type: object
      required: [data, total, page, limit, totalPages]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Musica"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer

    CatalogStats:
      type: object
      required: [totalSongs, totalArtists]
      properties:
        totalSongs:
          type: integer
        totalArtists:
          type: integer

    KaraokeUser:
      type: object
      required: [id, email, subscriptionStatus, accessGranted]
      properties:
        id:
          type: integer
        email:
          type: string
        woocommerceCustomerId:
          type: ["string", "null"]
        subscriptionStatus:
          type: string
          enum: [active, cancelled, expired]
        accessGranted:
          type: boolean
        accessGrantedAt:
          type: ["string", "null"]
        expiresAt:
          type: ["string", "null"]

    WooCommerceEvent:
      type: object
      required: [event]
      properties:
        event:
          type: string
        customer_id:
          type: ["string", "null"]
        customer_email:
          type: ["string", "null"]
        subscription_status:
          type: ["string", "null"]
        order_id:
          type: ["string", "null"]

    WebhookResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string