# lib/api-zod/src/generated/types/catalogStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Karaoke CT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogStats {
  totalSongs: number;
  totalArtists: number;
}