# lib/api-zod/src/generated/types/karaokeUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Karaoke CT API specification
 * OpenAPI spec version: 0.1.0
 */
import type { KaraokeUserSubscriptionStatus } from './karaokeUserSubscriptionStatus';

export interface KaraokeUser {
  id: number;
  email: string;
  /** @nullable */
  woocommerceCustomerId?: string | null;
  subscriptionStatus: KaraokeUserSubscriptionStatus;
  accessGranted: boolean;
  /** @nullable */
  accessGrantedAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}