# lib/api-zod/src/generated/types/musicaInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Karaoke CT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MusicaInput {
  id: number;
  /** @minLength 1 */
  artista: string;
  /** @minLength 1 */
  musica: string;
  inicio?: string;
}