# lib/api-zod/src/generated/types/musicaSearchResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Karaoke CT API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Musica } from './musica';

export interface MusicaSearchResult {
  data: Musica[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}