# lib/api-zod/src/generated/types/wooCommerceEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Karaoke CT API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WooCommerceEvent {
  event: string;
  /** @nullable */
  customer_id?: string | null;
  /** @nullable */
  customer_email?: string | null;
  /** @nullable */
  subscription_status?: string | null;
  /** @nullable */
  order_id?: string | null;
}